+++ CHANGELOG.md
@@ -1,12 +1,29 @@
-v1.0.1      - 7th April 2022 - Initial release
+v1.0.8 20th July 2023 Cache the token in a file, only generating a new one if the cached one is not avaliable or expired.
 
-v1.0.2      - 8th April 2022 - Update Documentation
+Firmware 5 and 7 compatible thanks to helderfmf
 
-v1.0.3      - 8th April 2022 - Added Mqtt Port to Configuration
+** Note - you must now configure your Envoy username and password to obtain token for Firmware 7
 
-v1.0.4      - 10th April 2022 - Added delay in start
+Auto find which firmware is installed in Envoy.
+
+Auto generate token for firmware 7.
+
+Auto generate password for firmware 5.
+
+MQTT_TOPIC configurable.
+
+MQTT_TOPIC_FREEDS configurable Optionally.
+
+Get values with 1s interval for FW 7.
+
+v1.0.7      - 22nd December 2022 - Updated Documentation for new mqtt integration method in configuration.yaml
 
 v1.0.5      - 12th April 2022 - Updated documentation url for online password generator
 
-v1.0.6      - 16th May 2022 - Updated to fix potential s6-overlay issue
+v1.0.4      - 10th April 2022 - Added delay in start
 
+v1.0.3      - 8th April 2022 - Added Mqtt Port to Configuration
+
+v1.0.2      - 8th April 2022 - Update Documentation
+
+v1.0.1      - 7th April 2022 - Initial release